requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
